requirements.txt
# Copilot: Generate dependencies list for "AI DermalScan".
# Include all libraries required for model training, face detection, and Streamlit UI.
# Example:
# Requirements for AI_DermalScan
# Core ML and data libraries
tensorflow>=2.11,<3
tensorflow-hub>=0.12
tensorflow-datasets>=4.8
keras>=2.11
numpy>=1.23

# Image processing and computer vision
opencv-python>=4.7
Pillow>=9.4
scikit-image>=0.20

# Data science and plotting
pandas>=1.5
matplotlib>=3.6
seaborn>=0.12
scikit-learn>=1.2

# Web UI
streamlit>=1.18

# Utilities
tqdm>=4.64

# Optional (useful during development)
jupyterlab
